moved swap and multiswap functions to separate package (#16)
* move cache stub functionality to separate package 'cache_stub'


* move cache stub functionality to separate package 'cache_stub'


* moved swap, multiswap functionality to separate packages 'swap' and 'multiswap'


* moved swap, multiswap functionality to separate packages 'swap' and 'multiswap'


* delete wrong files
- doc.go
- example_test.go
- mock_cache_stub_test.go


---------
