+++ core/multiswap.go
@@ -5,158 +5,46 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"log"
 	mathbig "math/big"
-	"runtime/debug"
-	"strings"
 
 	"github.com/anoideaopen/foundation/core/balance"
 	"github.com/anoideaopen/foundation/core/cachestub"
+	"github.com/anoideaopen/foundation/core/multiswap"
+	"github.com/anoideaopen/foundation/core/telemetry"
 	"github.com/anoideaopen/foundation/core/types"
 	"github.com/anoideaopen/foundation/core/types/big"
 	"github.com/anoideaopen/foundation/proto"
-	pb "github.com/golang/protobuf/proto" //nolint:staticcheck
 	"github.com/hyperledger/fabric-chaincode-go/shim"
 	"github.com/hyperledger/fabric-protos-go/peer"
-	"golang.org/x/crypto/sha3"
 )
 
-const (
-	// MultiSwapCompositeType is a composite key for multiswap
-	MultiSwapCompositeType = "multi_swap"
-	// MultiSwapKeyEvent is a reason for multiswap
-	MultiSwapKeyEvent = "multi_swap_key"
-)
-
-func multiSwapAnswer(stub *cachestub.BatchCacheStub, swap *proto.MultiSwap) (r *proto.SwapResponse) {
-	r = &proto.SwapResponse{Id: swap.Id, Error: &proto.ResponseError{Error: "panic multiSwapAnswer"}}
-	defer func() {
-		if rc := recover(); rc != nil {
-			log.Println("panic multiSwapAnswer: " + hex.EncodeToString(swap.Id) + "\n" + string(debug.Stack()))
-		}
-	}()
-
-	ts, err := stub.GetTxTimestamp()
-	if err != nil {
-		return &proto.SwapResponse{Id: swap.Id, Error: &proto.ResponseError{Error: err.Error()}}
-	}
-	txStub := stub.NewTxCacheStub(hex.EncodeToString(swap.Id))
-
-	swap.Creator = []byte("0000")
-	swap.Timeout = ts.Seconds + robotSideTimeout
-
-	switch {
-	case swap.Token == swap.From:
-		// nothing to do
-	case swap.Token == swap.To:
-		for _, asset := range swap.Assets {
-			if err = balance.Sub(txStub, balance.BalanceTypeGiven, swap.From, "", new(mathbig.Int).SetBytes(asset.Amount)); err != nil {
-				return &proto.SwapResponse{Id: swap.Id, Error: &proto.ResponseError{Error: err.Error()}}
-			}
-		}
-	default:
-		return &proto.SwapResponse{Id: swap.Id, Error: &proto.ResponseError{Error: ErrIncorrectSwap}}
+// multiSwapDoneHandler processes a request to mark multiple swaps as done.
+// If the ChainCode is configured to disable multi swaps, it will immediately return an error.
+//
+// It loads initial arguments and then proceeds to execute the multi-swap user done logic.
+//
+// Returns a shim.Success response if the multi-swap done logic executes successfully.
+// Otherwise, it returns a shim.Error response.
+func (cc *ChainCode) multiSwapDoneHandler(
+	traceCtx telemetry.TraceContext,
+	stub shim.ChaincodeStubInterface,
+	args []string,
+	cfgBytes []byte,
+) peer.Response {
+	if cc.contract.ContractConfig().Options.DisableMultiSwaps {
+		return shim.Error(fmt.Sprintf(
+			"handling multi-swap done failed, %s", ErrMultiSwapDisabled.Error(),
+		))
 	}
 
-	if err = MultiSwapSave(txStub, hex.EncodeToString(swap.Id), swap); err != nil {
-		return &proto.SwapResponse{Id: swap.Id, Error: &proto.ResponseError{Error: err.Error()}}
-	}
-	writes, _ := txStub.Commit()
-	return &proto.SwapResponse{Id: swap.Id, Writes: writes}
-}
+	_, contract := copyContractWithConfig(traceCtx, cc.contract, stub, cfgBytes)
 
-func multiSwapRobotDone(stub *cachestub.BatchCacheStub, swapID []byte, key string) (r *proto.SwapResponse) {
-	r = &proto.SwapResponse{Id: swapID, Error: &proto.ResponseError{Error: "panic multiSwapRobotDone"}}
-	defer func() {
-		if rc := recover(); rc != nil {
-			log.Println("panic multiSwapRobotDone: " + hex.EncodeToString(swapID) + "\n" + string(debug.Stack()))
-		}
-	}()
-
-	txStub := stub.NewTxCacheStub(hex.EncodeToString(swapID))
-	swap, err := MultiSwapLoad(txStub, hex.EncodeToString(swapID))
-	if err != nil {
-		return &proto.SwapResponse{Id: swapID, Error: &proto.ResponseError{Error: err.Error()}}
-	}
-	hash := sha3.Sum256([]byte(key))
-	if !bytes.Equal(swap.Hash, hash[:]) {
-		return &proto.SwapResponse{Id: swapID, Error: &proto.ResponseError{Error: ErrIncorrectKey}}
-	}
-
-	if swap.Token == swap.From {
-		for _, asset := range swap.Assets {
-			if err = balance.Add(txStub, balance.BalanceTypeGiven, swap.To, "", new(mathbig.Int).SetBytes(asset.Amount)); err != nil {
-				return &proto.SwapResponse{Id: swapID, Error: &proto.ResponseError{Error: err.Error()}}
-			}
-		}
-	}
-
-	if err = MultiSwapDel(txStub, hex.EncodeToString(swapID)); err != nil {
-		return &proto.SwapResponse{Id: swapID, Error: &proto.ResponseError{Error: err.Error()}}
-	}
-	writes, _ := txStub.Commit()
-	return &proto.SwapResponse{Id: swapID, Writes: writes}
-}
-
-func multiSwapUserDone(bc BaseContractInterface, swapID string, key string) peer.Response {
-	swap, err := MultiSwapLoad(bc.GetStub(), swapID)
-	if err != nil {
-		return shim.Error(err.Error())
-	}
-	hash := sha3.Sum256([]byte(key))
-	if !bytes.Equal(swap.Hash, hash[:]) {
-		return shim.Error(ErrIncorrectKey)
-	}
-
-	if bytes.Equal(swap.Creator, swap.Owner) {
-		return shim.Error(ErrIncorrectSwap)
-	}
-	if swap.Token == swap.From {
-		if err = bc.AllowedIndustrialBalanceAdd(types.AddrFromBytes(swap.Owner), swap.Assets, "multi-swap done"); err != nil {
-			return shim.Error(err.Error())
-		}
-	} else {
-		for _, asset := range swap.Assets {
-			if err = bc.TokenBalanceAddWithTicker(types.AddrFromBytes(swap.Owner), new(big.Int).SetBytes(asset.Amount), asset.Group, "reverse multi-swap done"); err != nil {
-				return shim.Error(err.Error())
-			}
-		}
-	}
-
-	if err = MultiSwapDel(bc.GetStub(), swapID); err != nil {
-		return shim.Error(err.Error())
-	}
-	e := strings.Join([]string{swap.From, swapID, key}, "\t")
-	if err = bc.GetStub().SetEvent(MultiSwapKeyEvent, []byte(e)); err != nil {
-		return shim.Error(err.Error())
-	}
-
-	// This code implements a callback which notifies that MultiSwap was made.
-	// This callback handles direct (move tokens to other channel) or
-	// reverse (move tokens from other channel back) MultiSwaps.
-	// If you want to catch that events you need implement
-	// method `OnMultiSwapDoneEvent` in chaincode.
-	// This code is for chaincode PFT, for handling user bar tokens balance changes.
-	if f, ok := bc.(interface {
-		OnMultiSwapDoneEvent(
-			token string,
-			owner *types.Address,
-			assets []*proto.Asset,
-		)
-	}); ok {
-		f.OnMultiSwapDoneEvent(
-			swap.Token,
-			types.AddrFromBytes(swap.Owner),
-			swap.Assets,
-		)
-	}
-
-	return shim.Success(nil)
+	return multiswap.UserDone(contract, args[0], args[1])
 }
 
 // QueryMultiSwapGet - returns multiswap by id
 func (bc *BaseContract) QueryMultiSwapGet(swapID string) (*proto.MultiSwap, error) {
-	swap, err := MultiSwapLoad(bc.GetStub(), swapID)
+	swap, err := multiswap.Load(bc.GetStub(), swapID)
 	if err != nil {
 		return nil, err
 	}
@@ -205,10 +93,10 @@ func (bc *BaseContract) TxMultiSwapBegin(sender *types.Sender, token string, mul
 			return "", err
 		}
 	default:
-		return "", errors.New(ErrIncorrectSwap)
+		return "", errors.New(multiswap.ErrIncorrectMultiSwap)
 	}
 
-	if err = MultiSwapSave(bc.GetStub(), bc.GetStub().GetTxID(), &swap); err != nil {
+	if err = multiswap.Save(bc.GetStub(), bc.GetStub().GetTxID(), &swap); err != nil {
 		return "", err
 	}
 
@@ -220,7 +108,7 @@ func (bc *BaseContract) TxMultiSwapBegin(sender *types.Sender, token string, mul
 
 // TxMultiSwapCancel - cancels multiswap
 func (bc *BaseContract) TxMultiSwapCancel(sender *types.Sender, swapID string) error {
-	swap, err := MultiSwapLoad(bc.GetStub(), swapID)
+	swap, err := multiswap.Load(bc.GetStub(), swapID)
 	if err != nil {
 		return err
 	}
@@ -256,49 +144,7 @@ func (bc *BaseContract) TxMultiSwapCancel(sender *types.Sender, swapID string) e
 		}
 	}
 
-	return MultiSwapDel(bc.GetStub(), swapID)
-}
-
-// MultiSwapLoad - loads multiswap from the ledger
-func MultiSwapLoad(stub shim.ChaincodeStubInterface, swapID string) (*proto.MultiSwap, error) {
-	key, err := stub.CreateCompositeKey(MultiSwapCompositeType, []string{swapID})
-	if err != nil {
-		return nil, err
-	}
-	data, err := stub.GetState(key)
-	if err != nil {
-		return nil, err
-	}
-	if data == nil {
-		return nil, errors.New("multiswap doesn't exist")
-	}
-	var swap proto.MultiSwap
-	if err = pb.Unmarshal(data, &swap); err != nil {
-		return nil, err
-	}
-	return &swap, nil
-}
-
-// MultiSwapSave - saves multiswap to the ledger
-func MultiSwapSave(stub shim.ChaincodeStubInterface, swapID string, swap *proto.MultiSwap) error {
-	key, err := stub.CreateCompositeKey(MultiSwapCompositeType, []string{swapID})
-	if err != nil {
-		return err
-	}
-	data, err := pb.Marshal(swap)
-	if err != nil {
-		return err
-	}
-	return stub.PutState(key, data)
-}
-
-// MultiSwapDel - deletes multiswap from the ledger
-func MultiSwapDel(stub shim.ChaincodeStubInterface, swapID string) error {
-	key, err := stub.CreateCompositeKey(MultiSwapCompositeType, []string{swapID})
-	if err != nil {
-		return err
-	}
-	return stub.DelState(key)
+	return multiswap.Delete(bc.GetStub(), swapID)
 }
 
 // QueryGroupBalanceOf - returns balance of the token for user address
